fix(supabase): make menu pgTAP suite pass
- Add forward migration to fix PL/pgSQL ambiguity in increment_menu_usage (Postgres treats RETURNS TABLE columns as vars)\n- Fix pgTAP type mismatch in usage limits test (bigint count vs int)\n- Make recipe idempotency test deterministic and premium-aware

+++ supabase/tests/0022_menu_usage_limits.test.sql
@@ -8,7 +8,7 @@ select tests.create_supabase_user('menu-usage-user'::text) as user_id \gset
 -- ensure no row exists initially
 select is(
   (select count(*) from public.menu_usage_counters where owner_id = :'user_id'),
-  0,
+  0::bigint,
   'usage counters start empty'
 );
